requirements.txt
gradio>=4.0.0
plotly>=5.0.0
pandas>=1.3.0
numpy>=1.21.0
pillow>=9.0.0
